Added zoom and color to country on map
