+++ Javascript2/src/App.jsx
@@ -1,16 +1,10 @@
 import { Routes, Route } from "react-router";
-import { useState, useEffect } from "react";
-import { getAllCountries, getCountryBySlug } from "./api/countriesApi";
-import "./styles/WeatherWidget.css";
 import Navbar from "./components/Navbar";
 import "./styles/Navbar.css";
-import CountryInfoDetails from "./pages/CountryInfoDetails";
-import CountryInfo from "./components/CountryInfo";
+import CountryInfoDetails from "./pages/CountryInfoPage.jsx";
 import { getWeather } from "./api/weatherApi";
-import WeatherWidget from "./components/WeatherWidget";
 import FlagQuizPage from "./pages/FlagQuizPage.jsx";
 import { useCountriesStore } from "./store/countriesStore.js";
-import MapView from "./components/CountryMaps.jsx";
 
 function App() {
   return (
@@ -24,4 +18,3 @@ function App() {
 }
 
 export default App;
-
